feature(admin/client): add to cart feature added

+++ src/app/api/admin/delete-product/route.js
@@ -1,4 +1,5 @@
 import connectToDB from "@/database";
+import AuthUser from "@/middleware/AuthUser";
 import Product from "@/models/product";
 import { NextResponse } from "next/server";
 
@@ -7,20 +8,30 @@ export const dynamic = "force-dynamic";
 export async function DELETE(req) {
   try {
     await connectToDB();
-    const { searchParams } = new URL(req.url);
-    const id = searchParams.get("id");
-    if (!id) {
-      return NextResponse.json({
-        success: false,
-        message: "Product id is required",
-      });
-    }
-    const deletedProduct = await Product.findByIdAndDelete(id);
-    if (deletedProduct) {
-      return NextResponse.json({
-        success: true,
-        message: "Product is deleted Successfully",
-      });
+
+    const isAuthUser = await AuthUser(req);
+
+    if (isAuthUser?.role === "admin") {
+      const { searchParams } = new URL(req.url);
+      const id = searchParams.get("id");
+      if (!id) {
+        return NextResponse.json({
+          success: false,
+          message: "Product id is required",
+        });
+      }
+      const deletedProduct = await Product.findByIdAndDelete(id);
+      if (deletedProduct) {
+        return NextResponse.json({
+          success: true,
+          message: "Product is deleted Successfully",
+        });
+      } else {
+        return NextResponse.json({
+          success: false,
+          message: "Failed to delete Product",
+        });
+      }
     } else {
       return NextResponse.json({
         success: false,
@@ -30,7 +41,7 @@ export async function DELETE(req) {
   } catch (error) {
     return NextResponse.json({
       success: false,
-      message: "Something went wrong ! Please try again later",
+      message: "Your are not Authenticated",
     });
   }
 }
